go.d: split s3check provider responsibilities
